Add a hack to ensure unload_handler works correctly on ST2
For some reason I don't understand the reload on ST2 destroys the pool before any attempt to run a process, but the same does not happen on ST3. This adds a hack to try and differentiate between the module being disabled (unload_module) and the plugin simply being reloaded (reload_module) as happens at start-up.